refactor: drop edge package
