+++ packages/cli/src/codemods/v0-9/edge-package-split.ts
@@ -1,14 +1,9 @@
 import { createTransformer } from "../utils/createTransformer";
 
-const reactEdgeExports: string[] = [
-  // Edge Runtime
-  "useEdgeRuntime",
-  "EdgeRuntimeOptions",
-  "EdgeModelAdapter",
-  "EdgeChatAdapter",
-  "EdgeRuntimeRequestOptions",
-  "createEdgeRuntimeAPI",
-  "getEdgeRuntimeResponse",
+// Chat runtime and converter exports moved to react-ai-sdk
+const reactAiSdkExports: string[] = [
+  // Chat Runtime (replaces edge runtime)
+  "useChatRuntime",
 
   // Core Types
   "CoreMessage",
@@ -24,10 +19,8 @@ const reactEdgeExports: string[] = [
   "fromCoreMessage",
   "toCoreMessages",
   "toCoreMessage",
-];
 
-// Language model converters to be moved to react-ai-sdk
-const reactAiSdkExports: string[] = [
+  // Language model converters
   "toLanguageModelMessages",
   "toLanguageModelTools",
   "fromLanguageModelMessages",
@@ -37,7 +30,6 @@ const reactAiSdkExports: string[] = [
 
 const migrateToEdgePackage = createTransformer(({ j, root, markAsChanged }) => {
   const sourcesToMigrate: string[] = ["@assistant-ui/react"];
-  const movedEdgeSpecifiers: any[] = [];
   const movedAiSdkSpecifiers: any[] = [];
   let lastMigratedImportPath: any = null;
 
@@ -49,12 +41,6 @@ const migrateToEdgePackage = createTransformer(({ j, root, markAsChanged }) => {
       const remainingSpecifiers: any[] = [];
       path.value.specifiers.forEach((specifier: any) => {
         if (
-          j.ImportSpecifier.check(specifier) &&
-          reactEdgeExports.includes(specifier.imported.name as string)
-        ) {
-          movedEdgeSpecifiers.push(specifier);
-          hadMigratedSpecifiers = true;
-        } else if (
           j.ImportSpecifier.check(specifier) &&
           reactAiSdkExports.includes(specifier.imported.name as string)
         ) {
@@ -76,43 +62,7 @@ const migrateToEdgePackage = createTransformer(({ j, root, markAsChanged }) => {
       }
     });
 
-  // Add imports for react-edge
-  if (movedEdgeSpecifiers.length > 0) {
-    const existingEdgeImport = root.find(j.ImportDeclaration, {
-      source: { value: "@assistant-ui/react-edge" },
-    });
-    if (existingEdgeImport.size() > 0) {
-      existingEdgeImport.forEach((path: any) => {
-        movedEdgeSpecifiers.forEach((specifier: any) => {
-          if (
-            !path.value.specifiers.some(
-              (s: any) => s.imported.name === specifier.imported.name,
-            )
-          ) {
-            path.value.specifiers.push(specifier);
-          }
-        });
-      });
-    } else {
-      const newImport = j.importDeclaration(
-        movedEdgeSpecifiers,
-        j.literal("@assistant-ui/react-edge"),
-      );
-      if (lastMigratedImportPath) {
-        j(lastMigratedImportPath).insertAfter(newImport);
-      } else {
-        const firstImport = root.find(j.ImportDeclaration).at(0);
-        if (firstImport.size() > 0) {
-          firstImport.insertBefore(newImport);
-        } else {
-          root.get().node.program.body.unshift(newImport);
-        }
-      }
-    }
-    markAsChanged();
-  }
-
-  // Add imports for react-ai-sdk
+  // Add imports for react-ai-sdk (all edge-related imports now go here)
   if (movedAiSdkSpecifiers.length > 0) {
     const existingAiSdkImport = root.find(j.ImportDeclaration, {
       source: { value: "@assistant-ui/react-ai-sdk" },
@@ -159,7 +109,7 @@ const migrateToEdgePackage = createTransformer(({ j, root, markAsChanged }) => {
       path.value.source = j.literal(
         sourceValue.replace(
           "@assistant-ui/react/",
-          "@assistant-ui/react-edge/",
+          "@assistant-ui/react-ai-sdk/",
         ),
       );
       markAsChanged();
@@ -169,7 +119,10 @@ const migrateToEdgePackage = createTransformer(({ j, root, markAsChanged }) => {
   // Migrate language model converter imports from react-edge to react-ai-sdk
   root.find(j.ImportDeclaration).forEach((path: any) => {
     const sourceValue: string = path.value.source.value;
-    if (sourceValue === "@assistant-ui/react-edge") {
+    if (
+      sourceValue === "@assistant-ui/react-edge" ||
+      sourceValue === "@assistant-ui/react"
+    ) {
       let hasLanguageModelConverters = false;
       const remainingSpecifiers: any[] = [];
       const aiSdkSpecifiers: any[] = [];
