+++ packages/cli/src/lib/install-edge-lib.ts
@@ -49,34 +49,38 @@ export default async function installEdgeLib(): Promise<void> {
   for (const file of files) {
     const fullPath = path.join(cwd, file);
     const content = fs.readFileSync(fullPath, "utf8");
-    if (content.includes("@assistant-ui/react-edge")) {
+    if (
+      content.includes("@assistant-ui/react-edge") ||
+      (content.includes("useChatRuntime") &&
+        content.includes("@assistant-ui/react-ai-sdk"))
+    ) {
       found = true;
       break;
     }
   }
 
   if (found) {
-    if (isPackageInstalled("@assistant-ui/react-edge")) {
+    if (isPackageInstalled("@assistant-ui/react-ai-sdk")) {
       console.log(
-        "@assistant-ui/react-edge is already installed. Skipping installation.",
+        "@assistant-ui/react-ai-sdk is already installed. Skipping installation.",
       );
       return;
     }
 
     const answer = await askQuestion(
-      "Edge Runtime imports were added but @assistant-ui/react-edge is not installed. Do you want to install it? (Y/n) ",
+      "Edge Runtime imports were detected but @assistant-ui/react-ai-sdk is not installed. Do you want to install it? (Y/n) ",
     );
     if (answer === "" || answer.toLowerCase().startsWith("y")) {
       const pm = await detect();
       let cmd = "";
       if (pm === "yarn") {
-        cmd = "yarn add @assistant-ui/react-edge";
+        cmd = "yarn add @assistant-ui/react-ai-sdk";
       } else if (pm === "pnpm") {
-        cmd = "pnpm add @assistant-ui/react-edge";
+        cmd = "pnpm add @assistant-ui/react-ai-sdk";
       } else if (pm === "bun") {
-        cmd = "bun add @assistant-ui/react-edge";
+        cmd = "bun add @assistant-ui/react-ai-sdk";
       } else {
-        cmd = "npm install @assistant-ui/react-edge";
+        cmd = "npm install @assistant-ui/react-ai-sdk";
       }
       try {
         execSync(cmd, { stdio: "inherit" });
