+++ packages/react-ai-sdk/package.json
@@ -22,12 +22,15 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@ai-sdk/provider": "^1.1.3",
     "@ai-sdk/react": "*",
     "@ai-sdk/ui-utils": "*",
-    "@assistant-ui/react-edge": "workspace:*",
     "@radix-ui/react-use-callback-ref": "^1.1.1",
     "@types/json-schema": "^7.0.15",
+    "assistant-stream": "^0.2.17",
+    "json-schema": "^0.4.0",
     "zod": "^3.25.67",
+    "zod-to-json-schema": "^3.24.6",
     "zustand": "^5.0.6"
   },
   "peerDependencies": {
@@ -41,6 +44,7 @@
     }
   },
   "devDependencies": {
+    "@assistant-ui/react": "workspace:*",
     "@assistant-ui/x-buildutils": "workspace:*",
     "@types/node": "^24.0.4",
     "@types/react": "^19.1.8",
