+++ packages/react-ai-sdk/src/converters/toLanguageModelTools.ts
