+++ packages/react-ai-sdk/src/dangerous-in-browser/index.ts
