+++ packages/react-ai-sdk/src/index.ts
@@ -2,12 +2,6 @@ export * from "./rsc";
 export * from "./ui";
 export * from "./useChatRuntime";
 export * from "./useCloudRuntime";
-
-export {
-  toLanguageModelMessages,
-  toLanguageModelTools,
-  fromLanguageModelMessages,
-  fromLanguageModelTools,
-  useDangerousInBrowserRuntime,
-} from "@assistant-ui/react-edge";
+export * from "./converters";
+export * from "./dangerous-in-browser";
 export { frontendTools } from "./frontendTools";
