--- packages/react-edge/CHANGELOG.md
@@ -1,178 +1,0 @@
-# @assistant-ui/react-edge
-
-## 0.2.13
-
-### Patch Changes
-
-- 65b3ff1: chore: update deps
-- Updated dependencies [65b3ff1]
-- Updated dependencies [2731323]
-- Updated dependencies [308afff]
-- Updated dependencies [cc9f567]
-- Updated dependencies [c380f37]
-  - assistant-stream@0.2.18
-  - @assistant-ui/react@0.10.25
-
-## 0.2.12
-
-### Patch Changes
-
-- 644abb8: chore: update deps
-- Updated dependencies [b65e354]
-- Updated dependencies [8eda24b]
-- Updated dependencies [644abb8]
-  - @assistant-ui/react@0.10.24
-  - assistant-stream@0.2.17
-
-## 0.2.11
-
-### Patch Changes
-
-- 3fd9ec1: fix: useAssistant should correctly display
-- Updated dependencies [3ed39ef]
-  - @assistant-ui/react@0.10.18
-
-## 0.2.10
-
-### Patch Changes
-
-- chore: update deps
-- Updated dependencies
-  - assistant-stream@0.2.10
-  - @assistant-ui/react@0.10.12
-
-## 0.2.9
-
-### Patch Changes
-
-- 1a4927a: feat: ThreadListItem.detach()
-- Updated dependencies [9811ff8]
-- Updated dependencies [1a4927a]
-  - @assistant-ui/react@0.10.10
-
-## 0.2.8
-
-### Patch Changes
-
-- b625527: feat: useChatRuntime onCancel option
-
-## 0.2.7
-
-### Patch Changes
-
-- 5cb9598: feat(assistant-stream): ObjectStream
-- Updated dependencies [5cb9598]
-  - assistant-stream@0.2.7
-  - @assistant-ui/react@0.10.8
-
-## 0.2.6
-
-### Patch Changes
-
-- a7551f8: feat: export EdgeRuntimeRequestOptionsSchema
-
-## 0.2.5
-
-### Patch Changes
-
-- 73a6ff1: feat: Tool.type
-- Updated dependencies [e9f8e7e]
-- Updated dependencies [c4c60cf]
-- Updated dependencies [73a6ff1]
-  - @assistant-ui/react@0.10.5
-  - assistant-stream@0.2.5
-
-## 0.2.4
-
-### Patch Changes
-
-- 98a680e: chore: update deps
-- Updated dependencies [98a680e]
-- Updated dependencies [98a680e]
-  - @assistant-ui/react@0.10.4
-  - assistant-stream@0.2.4
-
-## 0.2.3
-
-### Patch Changes
-
-- 30ae924: fix: disabled tools should still execute if invoked
-- Updated dependencies [30ae924]
-  - @assistant-ui/react@0.10.3
-
-## 0.2.2
-
-### Patch Changes
-
-- fix: ESM without bundler compat
-- Updated dependencies
-  - @assistant-ui/react@0.10.2
-
-## 0.2.1
-
-### Patch Changes
-
-- fix: correctly include Typescript declarations
-- Updated dependencies
-  - @assistant-ui/react@0.10.1
-
-## 0.2.0
-
-### Patch Changes
-
-- 557c3f7: build: drop CJS builds
-- Updated dependencies [557c3f7]
-  - @assistant-ui/react@0.9.7
-
-## 0.1.6
-
-### Patch Changes
-
-- chore: update deps
-- Updated dependencies
-  - @assistant-ui/react@0.9.6
-
-## 0.1.5
-
-### Patch Changes
-
-- 91897ee: fix: circular import causes zod schemas to break
-
-## 0.1.4
-
-### Patch Changes
-
-- chore: bump assistant-stream dependency
-- Updated dependencies
-- Updated dependencies [1ad0696]
-  - @assistant-ui/react@0.9.5
-
-## 0.1.3
-
-### Patch Changes
-
-- b9c731a: chore: update dependencies
-- Updated dependencies [62c2af7]
-- Updated dependencies [b9c731a]
-  - @assistant-ui/react@0.9.3
-
-## 0.1.2
-
-### Patch Changes
-
-- c0c9422: feat: useToolArgsFieldStatus
-- Updated dependencies [553bdff]
-- Updated dependencies [c0c9422]
-- Updated dependencies [675fb20]
-- Updated dependencies [4e86ab4]
-- Updated dependencies [e893985]
-- Updated dependencies [0500584]
-  - @assistant-ui/react@0.9.2
-
-## 0.1.1
-
-### Patch Changes
-
-- chore: update deps
-- Updated dependencies
-  - @assistant-ui/react@0.9.1
