--- packages/react-edge/eslint.config.ts
@@ -1,1 +1,0 @@
-export { default } from "@assistant-ui/x-buildutils/eslint";
