--- packages/react-edge/package.json
@@ -1,76 +1,0 @@
-{
-  "name": "@assistant-ui/react-edge",
-  "description": "The legacy edge runtime for assistant-ui",
-  "version": "0.2.13",
-  "license": "MIT",
-  "type": "module",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    },
-    "./edge": {
-      "types": "./dist/edge.d.ts",
-      "default": "./dist/edge.js"
-    }
-  },
-  "source": "./src/index.ts",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist",
-    "src",
-    "edge",
-    "README.md"
-  ],
-  "sideEffects": false,
-  "scripts": {
-    "build": "tsx scripts/build.mts",
-    "lint": "eslint ."
-  },
-  "dependencies": {
-    "@ai-sdk/provider": "^1.1.3",
-    "assistant-stream": "^0.2.18",
-    "json-schema": "^0.4.0",
-    "zod": "^3.25.67",
-    "zod-to-json-schema": "^3.24.6"
-  },
-  "peerDependencies": {
-    "@assistant-ui/react": "*",
-    "@types/react": "*",
-    "@types/react-dom": "*",
-    "react": "^18 || ^19 || ^19.0.0-rc",
-    "react-dom": "^18 || ^19 || ^19.0.0-rc"
-  },
-  "peerDependenciesMeta": {
-    "@types/react": {
-      "optional": true
-    },
-    "@types/react-dom": {
-      "optional": true
-    }
-  },
-  "devDependencies": {
-    "@assistant-ui/x-buildutils": "workspace:*",
-    "@assistant-ui/react": "workspace:*",
-    "@types/json-schema": "^7.0.15",
-    "@types/node": "^24.0.4",
-    "@types/react": "^19.1.8",
-    "@types/react-dom": "^19.1.6",
-    "eslint": "^9",
-    "eslint-config-next": "15.3.4",
-    "tsx": "^4.20.3"
-  },
-  "publishConfig": {
-    "access": "public",
-    "provenance": true
-  },
-  "homepage": "https://www.assistant-ui.com/",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/assistant-ui/assistant-ui/tree/main/packages/react-edge"
-  },
-  "bugs": {
-    "url": "https://github.com/assistant-ui/assistant-ui/issues"
-  }
-}
