--- packages/react-edge/scripts/build.mts
@@ -1,3 +1,0 @@
-import { Build } from "@assistant-ui/x-buildutils";
-
-await Build.start().transpileTypescript();
