--- packages/react-edge/src/edge/index.ts
@@ -1,12 +1,0 @@
-export * from "./CoreTypes";
-
-export { useEdgeRuntime, type EdgeRuntimeOptions } from "./useEdgeRuntime";
-export { EdgeModelAdapter as EdgeChatAdapter } from "./EdgeModelAdapter";
-export type { EdgeRuntimeRequestOptions } from "./EdgeRuntimeRequestOptions";
-
-export {
-  createEdgeRuntimeAPI,
-  getEdgeRuntimeResponse,
-} from "./createEdgeRuntimeAPI";
-
-export { EdgeRuntimeRequestOptionsSchema } from "./EdgeRuntimeRequestOptions";
