--- packages/react-edge/src/index.ts
@@ -1,3 +1,0 @@
-export * from "./dangerous-in-browser";
-export * from "./edge";
-export * from "./converters";
