--- packages/react-edge/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "@assistant-ui/x-buildutils/ts/base",
-  "include": ["src/**/*.ts", "src/**/*.tsx"],
-  "exclude": ["**/*.test.ts", "**/*.test.tsx"]
-}
