+++ packages/react/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "@assistant-ui/x-buildutils/ts/base",
   "include": ["src/**/*.ts", "src/**/*.tsx"],
-  "exclude": ["dist", "**/*.test.ts", "**/*.test.tsx"]
+  "exclude": ["**/*.test.ts", "**/*.test.tsx"]
 }
